requirements.txt
transformers
torch
faiss-cpu
accelerate